sbp fileio cleanup [ESD-1082] (#2765)
* Cleanup sbp_file_io module

* Differentiate seq number for read and write

* Make the internal seq number counters u32 instead of u8 similar
  to the SBP message declaration

* Add function docs

* Assert for too long filenames

* Add zero initializations

* Use strncpy to make clang-tidy happy, actual safety aspect is
  already covered by the added asserts.

* Unify error message for read and write

* Relocate ARM compiler version string printout

Currently version is output also for clean and other non-relevant
targets.

+++ src/sbp/sbp_fileio.h
@@ -17,13 +17,38 @@
 #include <swiftnav/common.h>
 
 void sbp_fileio_remove(const char *fn);
+
+/**
+ * @brief   Write to non-volatile memory
+ *
+ * @note    Function is neither reentrant nor thread-safe
+ *
+ * @param[in] filename      Name of the file to write to
+ * @param[in] offset        Offset into the file [bytes]
+ * @param[in] buf           Buffer to write from
+ * @param[in] size          Size of the buffer
+ *
+ * @return                  Number of bytes written
+ * @retval -1               Error
+ */
 ssize_t sbp_fileio_write(const char *filename,
-                         off_t offset,
+                         u32 offset,
                          const u8 *buf,
                          size_t size);
-ssize_t sbp_fileio_read(const char *filename,
-                        off_t offset,
-                        u8 *buf,
-                        size_t size);
+
+/**
+ * @brief   Read from non-volatile memory
+ *
+ * @note    Function is neither reentrant nor thread-safe
+ *
+ * @param[in] filename      Name of the file to read from
+ * @param[in] offset        Offset into the file [bytes]
+ * @param[in] buf           Buffer to write into
+ * @param[in] size          Size of the buffer
+ *
+ * @return                  Number of bytes read
+ * @retval -1               Error
+ */
+ssize_t sbp_fileio_read(const char *filename, u32 offset, u8 *buf, size_t size);
 
 #endif
